chore: re-org to include parent crate
