+++ rex-parser/src/lib.rs
@@ -1,4 +1,4 @@
-use std::{collections::VecDeque, fmt};
+use std::collections::VecDeque;
 
 use rex_ast::{
     ast::{Call, IfThenElse, Lambda, LetIn, Var, AST},
@@ -7,60 +7,14 @@ use rex_ast::{
 };
 
 use crate::{
+    error::{Error, ParserErr},
     lexer::{Token, Tokens},
     op::Operator,
 };
 
+pub mod error;
 pub mod lexer;
 pub mod op;
-
-#[derive(Clone, Debug, thiserror::Error)]
-pub enum Error {
-    #[error("parse errors {0:?}")]
-    Parser(Vec<ParserErr>),
-}
-
-impl From<Vec<ParserErr>> for Error {
-    fn from(errors: Vec<ParserErr>) -> Error {
-        Error::Parser(errors)
-    }
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct ParserErr {
-    span: Span,
-    message: String,
-}
-
-impl ParserErr {
-    pub fn new(span: Span, message: String) -> ParserErr {
-        ParserErr { span, message }
-    }
-}
-
-impl fmt::Display for ParserErr {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}: {}", self.span.begin, self.message)
-    }
-}
-
-impl From<String> for ParserErr {
-    fn from(message: String) -> ParserErr {
-        ParserErr {
-            span: Span::new(0, 0, 0, 0),
-            message,
-        }
-    }
-}
-
-impl From<&str> for ParserErr {
-    fn from(message: &str) -> ParserErr {
-        ParserErr {
-            span: Span::new(0, 0, 0, 0),
-            message: message.to_string(),
-        }
-    }
-}
 
 pub struct Parser {
     pub id_dispenser: IdDispenser,
