+++ rex/src/lib.rs
@@ -1,0 +1,4 @@
+pub use rex_ast::*;
+pub use rex_engine::*;
+pub use rex_parser::*;
+pub use rex_resolver::*;
